waf/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.28.1
redis==5.0.1
sqlalchemy==2.0.23
asyncpg==0.28.0
python-dotenv==1.0.0
pymongo==4.13.2
